[quant][be] Move DefaultFuseHandler to backend_config_utils.py
Summary:
att, just restructring code to make it clearer

Test Plan:
NA

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
